% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_plot.R
\name{histogram_plot}
\alias{histogram_plot}
\title{Make histogram plots}
\usage{
histogram_plot(data, group, variable, x_label)
}
\arguments{
\item{data}{The dataset}

\item{group}{The group variable}

\item{variable}{The variable for histograms}

\item{x_label}{The x_label for the plots.}
}
\value{
several histogram plots
}
\description{
This function make the histogram plots for a certain variable in different groups/levels seperately.
}
\details{
The function was made by combining "group_by", "mutate", and "ggplot" functions
}
\examples{
require(dplyr)
require(magrittr)
require(tidyverse)
require(datateachr)
require(gapminder)
require(glue)
histogram_plot(cancer_sample,diagnosis,radius_mean,"radius_mean")
histogram_plot(gapminder,continent,lifeExp,"lifeExp")

}
